Improve cross-compilation support for Windows
While there is no canonical capitalization for "windows.h" it has always
been all lowercase on implementations using case-sensitive file systems,
e.g. a cross-compiler running on Linux. Same goes for most (but not all)
other headers including "shlobj.h".

Ideally a configuration macro such as NOMINIMAX is defined before any
includes. Since it's so critical, libstdc++ on Windows automatically
defines it in its own headers if included first, which results in a
warning about a macro redefinition when compiling with GCC. Either move
the _WIN32 section before other includes, or as in this commit, detect
if it happened and avoid the redefinition.