Fix three data races with the indicator thread
The three "success" global variables are written by the main thread and
read by the indicator thread, constituting a data race. Simplest option
is to make them atomics, as no further synchronization is truly needed
to avoid a race condition.

A cast loads the values for passing to printf. This is an "int" cast
despite the underlying types because that is the type expected for the
format string (i.e. "%i"). The original types were incorrect, and the
compiler warning about this is suppressed due to indirection through
replaceColors().

There is still both a data race and race condition on output_length, to
be addressed separately.