Eliminate races and deadlock on indicator cleanup
There was a data race on output_length, a race condition on stderr, and
a deadlock in the signal handler:

(data race): The global variable output_length is written by the
indicator thread and read by the main thread without synchronization.

(race condition): Despite the stop request, the indicator thread
continues running concurrently. The output_length may continue to change
(i.e. it may be stale by the time it's used), and the indicator may even
print additional status output after cleanup because it has no yet seen
the stop request.

(deadlock): The signal handler uses stdio, which involves grabbing a
lock that may already be held by the thread handling the signal.

The naive fix is to make output_length atomic, eliminating the data
race, but leaving the race condition. However, to eliminate the race
condition, the main thread must wait to clean up the spinner until the
indicator thread has acknowledged the stop request. There's already a
tool for this: joining! So just use a regular thread.

The existing condition variable is already appropriate for requesting a
stop, as it provides a "wake-able sleep" so that it can be woken up
promptly for exiting. The old stop request boolean is converted to a
three-valued state. A canceled state is "sticky" and requests the
program to stop soon. If the indicator thread is running, printing the
cancelation message and exiting the program is its job. Otherwise the
signal handler will exit directly without a message.

Since the main thread is waiting for the indicator to exit, rather than
clean up after joining (i.e. multiple places), this responsibility is
left to the indicator thread.

Only the main thread mutates spinner_state, but the atomic exchanges int
{start,stop,cancel}Inidicator are still necessary because it races with
the signal handler.