Fix screen flicker on Wayland
Fixes #190

The GUIClipboardDaemon created a new Wayland window every ~2 seconds to
interact with the clipboard. This steals focus from the currently active
window for a very short time, causing context menus and overlays to
close. It basically made Clipboard unusable on Wayland.

The simplest strategy of fixing this is to align Linux with all other
platforms by only reading data from the clipboard on-demand instead of
doing it continuously.